Pieces and Possible Rotations added
Added the different tetris pieces and their possible rotations. I decided to use a 3d array to store the pieces as one and their possible rotations as a subset of each piece. The idea is to be able to isolate selecting the piece from performing a rotation on the piece. I already put the functionality in it's own file just to keep the game relatively modular. I plan to keep the pieces and piece selection separate from the game logic itself. I think this will create less bulky files and create more readability from file to file and also help me isolate where things go wrong.